duration are double in GRAS
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@2086 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ include/amok/bandwidth.h
@@ -85,10 +85,10 @@ double * amok_bw_matrix(xbt_dynar_t hosts, /* dynar of xbt_host_t */
 
 void amok_bw_saturate_start(const char* from_name,unsigned int from_port,
 			    const char* to_name,unsigned int to_port,
-			    unsigned int msg_size, unsigned int duration);
+			    unsigned int msg_size, double duration);
 
 void amok_bw_saturate_begin(const char* to_name,unsigned int to_port,
-			    unsigned int msg_size, unsigned int duration,
+			    unsigned int msg_size, double duration,
 			    /*out*/ double *elapsed, double *bw);
 
 void amok_bw_saturate_stop(const char* from_name,unsigned int from_port,
